fix: resolve audit issues — lazy load under spinlock, batch locks, WAL consistency
- Fix lazy load under spinlock: release per-node spinlock before
  bnode_entry_lazy_load_* (which does pread + mutex), re-acquire
  after I/O, re-find entry. New hbtrie_load_child_node() helper
  does I/O without writing to the entry, caller sets child pointer
  after re-acquiring lock. All 8 call sites in hbtrie_insert and
  hbtrie_delete fixed.

- Fix database_write_batch_sync: replace "acquire all 64 shard locks"
  with per-operation lock/unlock matching put_sync and delete_sync.
  Atomicity guaranteed by MVCC transaction, not shard locks.

- Fix WAL pending_writes tracking: increment for direct write path
  in DEBOUNCED/ASYNC modes, reset for ASYNC after buffer flush
  (data safe from process crash once in kernel buffer).

- Fix timer callback: reset pending_writes=0 for ASYNC mode after
  flush (prevents unbounded growth).

- Fix batch_count type: uint8_t -> uint16_t (wal_manager.h).

- Add defensive batch_size=1 enforcement for IMMEDIATE mode.

- Add ARM32 and PowerPC cpu_relax() hints in threadding.h.

- Remove dead two-phase write API (hbtrie_reserve/hbtrie_commit/
  hbtrie_reservation_destroy) — declared and implemented but never
  called; hbtrie_reserve just called hbtrie_insert synchronously.

- Add note on hbtrie_gc spinlock hold pattern.

+++ src/Database/wal_manager.h
@@ -92,7 +92,7 @@ typedef struct {
     // Batch write buffer (replaces debouncer)
     uint8_t entry_buf[4096];            // Pre-allocated entry buffer
     size_t entry_buf_used;              // Bytes used in entry_buf
-    uint8_t batch_count;                // Entries accumulated in current batch
+    uint16_t batch_count;                // Entries accumulated in current batch
     uint8_t batch_size;                 // Max entries before flush (1=IMMEDIATE, 0=buffer-full)
     int timer_active;                   // 1 if one-shot timer is pending
     uint64_t timer_id;                  // ID of the pending one-shot timer
